Fixed duplicated test
