+++ src/__mocks__/reconfigureControl.ts
@@ -1,0 +1,86 @@
+import { Control } from '../types';
+
+export const reconfigureControl = (
+  controlOverrides: Partial<Control> = {},
+): Control => ({
+  defaultValuesRef: {
+    current: {},
+  },
+  isWatchAllRef: {
+    current: false,
+  },
+  validFieldsRef: {
+    current: new Set(),
+  },
+  fieldsWithValidationRef: {
+    current: new Set(),
+  },
+  fieldArrayDefaultValues: {
+    current: {},
+  },
+  watchFieldsRef: {
+    current: new Set(),
+  },
+  dirtyFieldsRef: {
+    current: {},
+  },
+  watchFieldsHookRef: {
+    current: {},
+  },
+  watchFieldsHookRenderRef: {
+    current: {},
+  },
+  watchInternal: jest.fn(),
+  validateSchemaIsValid: jest.fn(),
+  getValues: jest.fn(),
+  reRender: jest.fn(),
+  setValue: jest.fn(),
+  register: jest.fn(),
+  unregister: jest.fn(),
+  trigger: jest.fn(),
+  removeFieldEventListener: jest.fn(),
+  errorsRef: { current: {} },
+  touchedFieldsRef: { current: {} },
+  mode: { isOnSubmit: false, isOnBlur: false, isOnChange: false },
+  reValidateMode: {
+    isReValidateOnBlur: false,
+    isReValidateOnSubmit: false,
+  },
+  formState: {
+    isDirty: false,
+    isSubmitted: false,
+    dirtyFields: {},
+    submitCount: 0,
+    touched: {},
+    isSubmitting: false,
+    isValid: false,
+  },
+  fieldsRef: {
+    current: {},
+  },
+  resetFieldArrayFunctionRef: {
+    current: {},
+  },
+  fieldArrayNamesRef: {
+    current: new Set<string>(),
+  },
+  isDirtyRef: {
+    current: false,
+  },
+  isSubmittedRef: {
+    current: false,
+  },
+  readFormStateRef: {
+    current: {
+      isDirty: true,
+      isSubmitted: false,
+      submitCount: false,
+      touched: false,
+      isSubmitting: false,
+      isValid: false,
+      dirtyFields: false,
+    },
+  },
+  renderWatchedInputs: () => {},
+  ...controlOverrides,
+});
